chore: refactor L3rdFloatingInspector, NoraFloatingInspector
